Refactor `IJsonFormatter<T, TSymbol>` Add EscapingHelper

+++ src/SpanJson/Helpers/DateTimeFormatter.Utf16.cs
@@ -9,7 +9,7 @@ namespace SpanJson.Helpers
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool TryFormat(DateTimeOffset value, Span<char> output, out int charsWritten)
         {
-            if (output.Length < 33)
+            if ((uint)output.Length < 33u)
             {
                 charsWritten = default;
                 return false;
@@ -33,7 +33,7 @@ namespace SpanJson.Helpers
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool TryFormat(DateTime value, Span<char> output, out int charsWritten)
         {
-            if (output.Length < 33)
+            if ((uint)output.Length < 33u)
             {
                 charsWritten = default;
                 return false;
@@ -57,7 +57,7 @@ namespace SpanJson.Helpers
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         private static void WriteDateAndTime(DateTime value, ref char c, out int charsWritten)
         {
-            WriteFourDigits((uint) value.Year, ref c, 0);
+            WriteFourDigits((uint)value.Year, ref c, 0);
             Unsafe.Add(ref c, 4) = '-';
             WriteTwoDigits(value.Month, ref c, 5);
             Unsafe.Add(ref c, 7) = '-';
@@ -69,7 +69,7 @@ namespace SpanJson.Helpers
             Unsafe.Add(ref c, 16) = ':';
             WriteTwoDigits(value.Second, ref c, 17);
             charsWritten = 19;
-            var fraction = (uint) ((ulong) value.Ticks % TimeSpan.TicksPerSecond);
+            var fraction = (uint)((ulong)value.Ticks % TimeSpan.TicksPerSecond);
             if (fraction > 0)
             {
                 Unsafe.Add(ref c, 19) = '.';
@@ -103,18 +103,18 @@ namespace SpanJson.Helpers
         private static void WriteFourDigits(uint value, ref char c, int startIndex)
         {
             var temp = '0' + value;
-            value /= 10;
-            Unsafe.Add(ref c, startIndex + 3) = (char) (temp - value * 10);
+            value /= 10u;
+            Unsafe.Add(ref c, startIndex + 3) = (char)(temp - value * 10u);
 
             temp = '0' + value;
-            value /= 10;
-            Unsafe.Add(ref c, startIndex + 2) = (char) (temp - value * 10);
+            value /= 10u;
+            Unsafe.Add(ref c, startIndex + 2) = (char)(temp - value * 10u);
 
             temp = '0' + value;
-            value /= 10;
-            Unsafe.Add(ref c, startIndex + 1) = (char) (temp - value * 10);
+            value /= 10u;
+            Unsafe.Add(ref c, startIndex + 1) = (char)(temp - value * 10u);
 
-            Unsafe.Add(ref c, startIndex) = (char) ('0' + value);
+            Unsafe.Add(ref c, startIndex) = (char)('0' + value);
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
@@ -122,8 +122,8 @@ namespace SpanJson.Helpers
         {
             var temp = '0' + value;
             value /= 10;
-            Unsafe.Add(ref c, startIndex + 1) = (char) (temp - value * 10);
-            Unsafe.Add(ref c, startIndex) = (char) ('0' + value);
+            Unsafe.Add(ref c, startIndex + 1) = (char)(temp - value * 10);
+            Unsafe.Add(ref c, startIndex) = (char)('0' + value);
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
@@ -132,8 +132,8 @@ namespace SpanJson.Helpers
             for (var i = 7; i > 0; i--)
             {
                 ulong temp = '0' + value;
-                value /= 10;
-                Unsafe.Add(ref c, pos + i - 1) = (char) (temp - value * 10);
+                value /= 10u;
+                Unsafe.Add(ref c, pos + i - 1) = (char)(temp - value * 10u);
             }
         }
     }
